refactor: 리소스 분리 및 공백 수정
